AG-11327 Move cartesian pickFocus into dataModelSeries
This allows the focus logic to be reused by polar and map series.
Currently, polar and map series have stubs for computeFocusBounds.
