+++ packages/ag-charts-enterprise/src/series/map-line-background/mapLineBackgroundSeries.ts
@@ -32,6 +32,10 @@ export class MapLineBackgroundSeries
     @Validate(GEOJSON_OBJECT, { optional: true, property: 'topology' })
     private _chartTopology?: _ModuleSupport.FeatureCollection = undefined;
 
+    protected override getNodeData(): MapLineBackgroundNodeDatum[] | undefined {
+        return this.contextNodeData?.nodeData;
+    }
+
     private get topology() {
         return this.properties.topology ?? this._chartTopology;
     }
@@ -196,5 +200,9 @@ export class MapLineBackgroundSeries
 
     override getTooltipHtml(): _ModuleSupport.TooltipContent {
         return _ModuleSupport.EMPTY_TOOLTIP_CONTENT;
+    }
+
+    protected override computeFocusBounds(_opts: _ModuleSupport.PickFocusInputs): _Scene.BBox | undefined {
+        return undefined; // TODO
     }
 }
