+++ packages/ag-charts-enterprise/src/series/map-line/mapLineSeries.ts
@@ -37,6 +37,10 @@ export class MapLineSeries
     @Validate(GEOJSON_OBJECT, { optional: true, property: 'topology' })
     private _chartTopology?: _ModuleSupport.FeatureCollection = undefined;
 
+    protected override getNodeData(): MapLineNodeDatum[] | undefined {
+        return this.contextNodeData?.nodeData;
+    }
+
     private get topology() {
         return this.properties.topology ?? this._chartTopology;
     }
@@ -629,5 +633,9 @@ export class MapLineSeries
                 ...this.getModuleTooltipParams(),
             }
         );
+    }
+
+    protected override computeFocusBounds(_opts: _ModuleSupport.PickFocusInputs): _Scene.BBox | undefined {
+        return undefined; // TODO
     }
 }
